+++ EasyPeasy/Attribute.swift
@@ -64,7 +64,25 @@ public class Attribute {
     /// Element identifying the node this attribute will be 
     /// stored in
     internal lazy var signature: String = {
-        return String.easy_signature(for: self)
+        // Signature of the create `ReferenceAttribute` of
+        // the passed `Attribute`
+        var signature = self.createAttribute.signatureString
+        
+        // Signature of the `Relation` of the passed `Attribute`
+        switch self.constant.relation {
+        case .Equal:
+            signature += "eq_"
+        case .GreaterThanOrEqual:
+            signature += "gt_"
+        case .LessThanOrEqual:
+            signature += "lt_"
+        }
+        
+        // Signature of the `Priority` of the passed
+        // `Attribute`
+        signature += String(self.priority.layoutPriority())
+        
+        return signature
     }()
     
     /**
@@ -74,7 +92,7 @@ public class Attribute {
         - returns: the `Attribute` instance created
      */
     public init() {
-        self.constant = Constant(0.0)
+        self.constant = Constant(value: 0.0, relation: .Equal, multiplier: 1.0)
         self.priority = .HighPriority
     }
     
@@ -86,7 +104,7 @@ public class Attribute {
         - returns: the `Attribute` instance created
      */
     public init(_ value: CGFloat) {
-        self.constant = Constant(value)
+        self.constant = Constant(value: value, relation: .Equal, multiplier: 1.0)
         self.priority = .HighPriority
     }
     
@@ -152,11 +170,11 @@ public class Attribute {
         let layoutConstraint = NSLayoutConstraint(
             item: item,
             attribute: self.createAttribute.layoutAttribute,
-            relatedBy: self.constant.layoutRelation(),
+            relatedBy: self.constant.relation,
             toItem: self.referenceItem,
             attribute: self.referenceAttributeHelper().layoutAttribute,
-            multiplier: self.constant.layoutMultiplier(),
-            constant: (self.constant.layoutValue() * constantFactor)
+            multiplier: self.constant.multiplier,
+            constant: (self.constant.value * constantFactor)
         )
         
         // Set priority
@@ -204,8 +222,8 @@ public class Attribute {
 }
 
 /**
-    Methods applicable to an `Array` of `Attributes`. The modifiers
-    will affect to each one of the `Attributes` within the `Array`
+    Methods applicable to an `Array` of `Attributes`. The `Constants`
+    will apply to each one of the `Attributes` within the `Array`
     overriding the values individually set.
  */
 public extension Array where Element: Attribute {