+++ Example/Tests/CompoundAttributeTests.swift
@@ -53,8 +53,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in sizeAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 0.0)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -66,8 +66,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in sizeAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -79,8 +79,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in sizeAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .GreaterThanOrEqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .GreaterThanOrEqual)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -92,8 +92,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in sizeAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 2.0)
-            XCTAssertTrue(attribute.constant.modifier == .MultipliedBy)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 2.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 2.0)
         }
     }
     
@@ -104,12 +104,12 @@ class CompoundAttributeTests: XCTestCase {
         
         // then
         XCTAssertTrue(sizeAttribute.attributes[0].constant.value == 120.0)
-        XCTAssertTrue(sizeAttribute.attributes[0].constant.modifier == .EqualTo)
-        XCTAssertTrue(sizeAttribute.attributes[0].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(sizeAttribute.attributes[0].constant.relation == .Equal)
+        XCTAssertTrue(sizeAttribute.attributes[0].constant.multiplier == 1.0)
         
         XCTAssertTrue(sizeAttribute.attributes[1].constant.value == 200.0)
-        XCTAssertTrue(sizeAttribute.attributes[1].constant.modifier == .EqualTo)
-        XCTAssertTrue(sizeAttribute.attributes[1].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(sizeAttribute.attributes[1].constant.relation == .Equal)
+        XCTAssertTrue(sizeAttribute.attributes[1].constant.multiplier == 1.0)
     }
     
     func testThatLikeMethodSetsTheReferenceViewToEachOneOfTheSubAttributes() {
@@ -135,8 +135,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in edgesAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 0.0)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -148,8 +148,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in edgesAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -161,8 +161,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in edgesAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .GreaterThanOrEqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .GreaterThanOrEqual)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -174,8 +174,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in edgesAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 2.0)
-            XCTAssertTrue(attribute.constant.modifier == .MultipliedBy)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 2.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 2.0)
         }
     }
     
@@ -186,20 +186,20 @@ class CompoundAttributeTests: XCTestCase {
         
         // then
         XCTAssertTrue(edgesAttribute.attributes[0].constant.value == 100.0)
-        XCTAssertTrue(edgesAttribute.attributes[0].constant.modifier == .EqualTo)
-        XCTAssertTrue(edgesAttribute.attributes[0].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(edgesAttribute.attributes[0].constant.relation == .Equal)
+        XCTAssertTrue(edgesAttribute.attributes[0].constant.multiplier == 1.0)
         
         XCTAssertTrue(edgesAttribute.attributes[1].constant.value == 20.0)
-        XCTAssertTrue(edgesAttribute.attributes[1].constant.modifier == .EqualTo)
-        XCTAssertTrue(edgesAttribute.attributes[1].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(edgesAttribute.attributes[1].constant.relation == .Equal)
+        XCTAssertTrue(edgesAttribute.attributes[1].constant.multiplier == 1.0)
         
         XCTAssertTrue(edgesAttribute.attributes[2].constant.value == 10.0)
-        XCTAssertTrue(edgesAttribute.attributes[2].constant.modifier == .EqualTo)
-        XCTAssertTrue(edgesAttribute.attributes[2].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(edgesAttribute.attributes[2].constant.relation == .Equal)
+        XCTAssertTrue(edgesAttribute.attributes[2].constant.multiplier == 1.0)
         
         XCTAssertTrue(edgesAttribute.attributes[3].constant.value == 50.0)
-        XCTAssertTrue(edgesAttribute.attributes[3].constant.modifier == .EqualTo)
-        XCTAssertTrue(edgesAttribute.attributes[3].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(edgesAttribute.attributes[3].constant.relation == .Equal)
+        XCTAssertTrue(edgesAttribute.attributes[3].constant.multiplier == 1.0)
     }
     
     func testThatCenterAttributeInitWithNoParamsSetsTheExpectedConstants() {
@@ -210,8 +210,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 0.0)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -223,8 +223,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -236,8 +236,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .GreaterThanOrEqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .GreaterThanOrEqual)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -249,8 +249,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 2.0)
-            XCTAssertTrue(attribute.constant.modifier == .MultipliedBy)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 2.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 2.0)
         }
     }
     
@@ -261,12 +261,12 @@ class CompoundAttributeTests: XCTestCase {
         
         // then
         XCTAssertTrue(centerAttribute.attributes[0].constant.value == 210.0)
-        XCTAssertTrue(centerAttribute.attributes[0].constant.modifier == .EqualTo)
-        XCTAssertTrue(centerAttribute.attributes[0].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(centerAttribute.attributes[0].constant.relation == .Equal)
+        XCTAssertTrue(centerAttribute.attributes[0].constant.multiplier == 1.0)
         
         XCTAssertTrue(centerAttribute.attributes[1].constant.value == 430.0)
-        XCTAssertTrue(centerAttribute.attributes[1].constant.modifier == .EqualTo)
-        XCTAssertTrue(centerAttribute.attributes[1].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(centerAttribute.attributes[1].constant.relation == .Equal)
+        XCTAssertTrue(centerAttribute.attributes[1].constant.multiplier == 1.0)
     }
     
     func testThatCenterWithinMarginsAttributeInitWithNoParamsSetsTheExpectedConstants() {
@@ -277,8 +277,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerWithinMarginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 0.0)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -290,8 +290,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerWithinMarginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -303,8 +303,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerWithinMarginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .GreaterThanOrEqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .GreaterThanOrEqual)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -316,8 +316,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in centerWithinMarginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 2.0)
-            XCTAssertTrue(attribute.constant.modifier == .MultipliedBy)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 2.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 2.0)
         }
     }
     
@@ -328,12 +328,12 @@ class CompoundAttributeTests: XCTestCase {
         
         // then
         XCTAssertTrue(centerWithinMarginsAttribute.attributes[0].constant.value == 210.0)
-        XCTAssertTrue(centerWithinMarginsAttribute.attributes[0].constant.modifier == .EqualTo)
-        XCTAssertTrue(centerWithinMarginsAttribute.attributes[0].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(centerWithinMarginsAttribute.attributes[0].constant.relation == .Equal)
+        XCTAssertTrue(centerWithinMarginsAttribute.attributes[0].constant.multiplier == 1.0)
         
         XCTAssertTrue(centerWithinMarginsAttribute.attributes[1].constant.value == 430.0)
-        XCTAssertTrue(centerWithinMarginsAttribute.attributes[1].constant.modifier == .EqualTo)
-        XCTAssertTrue(centerWithinMarginsAttribute.attributes[1].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(centerWithinMarginsAttribute.attributes[1].constant.relation == .Equal)
+        XCTAssertTrue(centerWithinMarginsAttribute.attributes[1].constant.multiplier == 1.0)
     }
     
     func testThatMaringsAttributeInitWithNoParamsSetsTheExpectedConstants() {
@@ -344,8 +344,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in marginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 0.0)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -357,8 +357,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in marginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .EqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -370,8 +370,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in marginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 50)
-            XCTAssertTrue(attribute.constant.modifier == .GreaterThanOrEqualTo)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 1.0)
+            XCTAssertTrue(attribute.constant.relation == .GreaterThanOrEqual)
+            XCTAssertTrue(attribute.constant.multiplier == 1.0)
         }
     }
     
@@ -383,8 +383,8 @@ class CompoundAttributeTests: XCTestCase {
         // then
         for attribute in marginsAttribute.attributes {
             XCTAssertTrue(attribute.constant.value == 2.0)
-            XCTAssertTrue(attribute.constant.modifier == .MultipliedBy)
-            XCTAssertTrue(attribute.constant.layoutMultiplier() == 2.0)
+            XCTAssertTrue(attribute.constant.relation == .Equal)
+            XCTAssertTrue(attribute.constant.multiplier == 2.0)
         }
     }
     
@@ -395,20 +395,20 @@ class CompoundAttributeTests: XCTestCase {
         
         // then
         XCTAssertTrue(marginsAttribute.attributes[0].constant.value == 100.0)
-        XCTAssertTrue(marginsAttribute.attributes[0].constant.modifier == .EqualTo)
-        XCTAssertTrue(marginsAttribute.attributes[0].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(marginsAttribute.attributes[0].constant.relation == .Equal)
+        XCTAssertTrue(marginsAttribute.attributes[0].constant.multiplier == 1.0)
         
         XCTAssertTrue(marginsAttribute.attributes[1].constant.value == 20.0)
-        XCTAssertTrue(marginsAttribute.attributes[1].constant.modifier == .EqualTo)
-        XCTAssertTrue(marginsAttribute.attributes[1].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(marginsAttribute.attributes[1].constant.relation == .Equal)
+        XCTAssertTrue(marginsAttribute.attributes[1].constant.multiplier == 1.0)
         
         XCTAssertTrue(marginsAttribute.attributes[2].constant.value == 10.0)
-        XCTAssertTrue(marginsAttribute.attributes[2].constant.modifier == .EqualTo)
-        XCTAssertTrue(marginsAttribute.attributes[2].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(marginsAttribute.attributes[2].constant.relation == .Equal)
+        XCTAssertTrue(marginsAttribute.attributes[2].constant.multiplier == 1.0)
         
         XCTAssertTrue(marginsAttribute.attributes[3].constant.value == 50.0)
-        XCTAssertTrue(marginsAttribute.attributes[3].constant.modifier == .EqualTo)
-        XCTAssertTrue(marginsAttribute.attributes[3].constant.layoutMultiplier() == 1.0)
+        XCTAssertTrue(marginsAttribute.attributes[3].constant.relation == .Equal)
+        XCTAssertTrue(marginsAttribute.attributes[3].constant.multiplier == 1.0)
     }
     
     func testThatCreateAtrributeIsWidthByDefault() {